+++ packages-integrations/vite/package.json
@@ -60,6 +60,7 @@
     "@unocss/inspector": "workspace:*",
     "chokidar": "catalog:",
     "magic-string": "catalog:",
+    "pathe": "catalog:",
     "tinyglobby": "catalog:",
     "unplugin-utils": "catalog:"
   },
