+++ packages-integrations/vite/src/modes/global/dev.ts
@@ -2,8 +2,9 @@ import type { GenerateResult, UnocssPluginContext } from '@unocss/core'
 import type { Plugin, Update, ViteDevServer } from 'vite'
 import type { VitePluginConfig } from '../../types'
 import process from 'node:process'
+import { LAYER_MARK_ALL } from '#integration/constants'
 import { getHash } from '#integration/hash'
-import { LAYER_MARK_ALL, resolveId, resolveLayer } from '#integration/layers'
+import { resolveId, resolveLayer } from '#integration/layers'
 import { getPath } from '#integration/utils'
 import { notNull } from '@unocss/core'
 import MagicString from 'magic-string'
